chore(sdk): scope agent sdk types package
