+++ package.json
@@ -199,17 +199,17 @@
     }
   },
   "scripts": {
-    "build": "npm run build -w agent-sdk-ts && npm run compile",
+    "build": "npm run build -w @openhands/agent-sdk-ts && npm run compile",
     "build:tailwind": "tailwindcss -i ./src/webview-src/tailwind.css -o ./src/webview-src/tailwind.gen.css --minify",
     "compile": "tsc -p . && npm run build:tailwind && npm run build:webview",
     "watch": "tsc -w -p . & npx tailwindcss -i ./src/webview-src/tailwind.css -o ./src/webview-src/tailwind.gen.css --watch & node --watch esbuild.webview.mjs",
     "vscode:prepublish": "npm run build",
     "package": "node scripts/run-vsce-package.js",
     "e2e": "npm run compile && tsc -p tests/e2e/tsconfig.suite.json && tsc -p tsconfig.e2e.json && mocha tests/e2e/out/**/*.test.js",
-    "test": "npm test -w agent-sdk-ts && vitest run --dir src",
+    "test": "npm test -w @openhands/agent-sdk-ts && vitest run --dir src",
     "test:watch": "vitest",
     "typecheck": "tsc -p . && tsc -p tsconfig.webview.json",
-    "lint": "npm run lint -w agent-sdk-ts && eslint src tests",
+    "lint": "npm run lint -w @openhands/agent-sdk-ts && eslint src tests",
     "lint:fix": "eslint . --fix",
     "build:webview": "node esbuild.webview.mjs",
     "dev:vscode": "bash scripts/dev-vscode.sh"
@@ -250,7 +250,7 @@
     "vitest": "^2.1.4"
   },
   "dependencies": {
-    "agent-sdk-ts": "workspace:*",
+    "@openhands/agent-sdk-ts": "workspace:*",
     "@openhands/ui": "^1.0.0-beta.9",
     "@types/react": "^19.1.15",
     "@types/react-dom": "^19.1.9",
