+++ packages/agent-sdk-ts/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "agent-sdk-ts",
+  "name": "@openhands/agent-sdk-ts",
   "version": "0.1.0",
   "description": "TypeScript models and guards for the OpenHands agent-sdk wire protocol",
   "license": "MIT",
